Change Map to LinkedHashMap to preserve insertion order
In the regular Map, the order of retreival is random. So, if you put in a,b,c in to the Map you might get b,c,a when you query them back. 
But that won't work for us as we need the order of lines in files intact. So, switching to LinkedHashMap. 
Unfortunately or Fortunately, it is mutable.