Add dedicated files for each delorean command
Since we are already doing it for some of the commands, its better to do it for all the commands. So, each command gets a class in "delorean.commands" package. 
So, add has Add class, status has Status class and so on. 
And it keep it even easier on the eye, each class has a companion object which will create the required object of the class. 
With this, It would be cleaner and easier to organize things even if more things need to be done for a command.