Add support for staging binary files as well
Now we can stage binary files as well like jars, zips etc.
Here we can't obviously get the file lines the way we usually get for other files instead we just get all the bytes of the file and hash that. 
Also we are  creating a new directory 'bins' for any binary files.