Fix the temp file not getting deleted issue
Temp file should be deleted once the pitstop has been made. But because of some resource not getting closed, it is not getting deleted. 
So, we are makeing sure we delete the file, first thing next time. 
We do this by setting the lastModifiedTime of the pitstop file  to be 10 more than the modified time of the temp file and we use that to delete it later.