Add 'create-timeline' command to create new timelines
Now we can create new timelines with
'delorean create-timeline' or 'delorean ctl'. 
Also added a test for this but right now it is kept uncommented until the time we have reset to a pitstop/timeline is ready.
This is because we are trying to check if the current timeline is "master" after we have told to switch to master. But the problem is that if the timeline master already exists and we are on a different timeline when we run the test, it won't switch to master. 
This can be possible only when resetting to a pistop or timeline is ready. So, for now just keeping the test under wraps.