Fix bug where partial pitstop hash was getting stored in the current indicator
We need to store the entire pistop hash in the "current" indicator file.
Before this what we were doing was that if the user does a "goto" with a partial
hash like "54asd3" then that is directly getting written to the current
indicator file which is not what we want.
We want the full hash for that pitstop "54asd3awqiojqweqwe......21" to be
added in the indicator.