Fix issue with FileReconstruction for binary files
For binary files, since the hash file won't be present under "hashes" directory but insteaf is under "bins",
the program was throwing an exception saying the file doesn't exist.
Fixed that by first checking if that file is a binary file or not and then progress accordingly.
If its a normal file, we'll do the standard reconstruction by reading from the string pool but for binary files, we just copy that file exactly.