Modify GoTo to use methods instead of putting everything in the constructor
Since I needed a way to exit for the case when the given "timeLine" is neither an actual timeline nor a pitstop, I split the code into methods.
If it is a timeline it'll fetch the pitstop corresponding to the timeline and then call the pitstop method which will do all the file operations needed.