Only take files that are not directories in untracked files
We don't want directories in the list of untracked files. If there is even one file in a directory that will be showed. Otherwise we'll just ignore it. This will make sure we don't show directories as untracked even when the files inside them were all staged.

But have to look for a better solution to this. Should decide whether we can ignore empty directories or not.