Add a directory parameter to Ride class
With this we don't have to actually be inside a directory to initiate a repo. We can be outside and call it.
Only use for this behaviour at this point is when we are starting the git server. Since we run it from the parent directory, we need "Ride"
to be able to create directories without that being the current directory.

May be later we can make "delorean ride" take an argument as well.